fix(presentation): prevent restore-on-upload logic from disabling itself
There is a mechanism to hide the presentation area while the
default presentation file is still uploading, to prevent displaying
an empty area during that time.

This mechanism lives inside the presentation component itself,
which gets unmounted when the presentation is hidden. As a result,
the mechanism disables itself and is no longer active to reopen the
presentation when the upload finishes. This only occurs when
`public.presentation.restoreOnUpdate` is set to false — when true,
the presentation is kept visible regardless, which masks the issue.

Move the restore-on-upload logic to the layout observer, in order
to prevent it from disabling itself.
